IGNITE-17539 Move collection of registered error groups to ErrorGroups
