+++ modules/transactions/src/main/java/org/apache/ignite/internal/tx/TransactionInternalCheckedException.java
@@ -17,9 +17,9 @@
 
 package org.apache.ignite.internal.tx;
 
-import static org.apache.ignite.lang.ErrorGroup.errorGroupByCode;
-import static org.apache.ignite.lang.ErrorGroup.extractGroupCode;
 import static org.apache.ignite.lang.ErrorGroups.Transactions.TX_ERR_GROUP;
+import static org.apache.ignite.lang.ErrorGroups.errorGroupByCode;
+import static org.apache.ignite.lang.ErrorGroups.extractGroupCode;
 
 import java.util.UUID;
 import org.apache.ignite.lang.IgniteInternalCheckedException;
